# server/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateBody": {
        "properties": {
          "name": {
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "type": "object"
      },
      "ErrorResponse": {
        "description": "Body of every non-2xx response.",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
      "Playlist": {
        "properties": {
          "icon": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "smart_rules": {
            "type": [
              "string",
              "null"
            ]
          },
          "track_count": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "name",
          "track_count"
        ],
        "type": "object"
      },
...
      "ReorderBody": {
        "properties": {
          "ids": {
            "items": {
              "format": "int64",
              "type": "integer"
            },
            "type": "array"
          }
        },
        "required": [
          "ids"
        ],
        "type": "object"
      },
...
      "SmartCreateBody": {
        "properties": {
          "name": {
            "type": "string"
          },
          "rules_json": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "rules_json"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "description": "The server's configured `api_key`.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "The Muorg music server HTTP API, shared by the desktop (Tauri), web and Android clients. Everything except `GET /api/health` and `GET /stream/{id}` requires an `Authorization: Bearer <api_key>` header.\n\n`GET /api/tracks` is paginated and defaults to 500 rows — a bare call returns the first page, not the catalog. Follow `X-Total-Count`.",
    "license": {
      "name": ""
    },
    "title": "Muorg Server API",
    "version": "2.42.1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/playlists": {
      "get": {
        "operationId": "list",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Playlist"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Playlists"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or invalid API key"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "All playlists, in user-defined order.",
        "tags": [
          "Playlists"
        ]
      },
      "post": {
        "operationId": "create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Playlist"
                }
              }
            },
            "description": "The created playlist"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or invalid API key"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Create an empty playlist.",
        "tags": [
          "Playlists"
        ]
      }
    },
    "/api/playlists/order": {
      "put": {
        "operationId": "reorder",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReorderBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            },
            "description": "Reordered"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or invalid API key"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Reorder the playlists themselves.",
        "tags": [
          "Playlists"
        ]
      }
    },
    "/api/playlists/smart": {
      "post": {
        "operationId": "create_smart",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SmartCreateBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Playlist"
                }
              }
            },
            "description": "The created playlist"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or invalid API key"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Create a rule-driven smart playlist.",
        "tags": [
          "Playlists"
        ]
      }
    },
    "/api/playlists/smart/{id}/rules": {
      "patch": {
        "operationId": "update_smart_rules",
        "parameters": [
          {
            "description": "Playlist id",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SmartRulesBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            },
            "description": "Updated"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or invalid API key"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "No such playlist"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Replace a smart playlist's rule set.",
        "tags": [
          "Playlists"
        ]
      }
    },
    "/api/playlists/smart/{id}/tracks": {
      "get": {
        "operationId": "get_smart_tracks",
        "parameters": [
          {
            "description": "Playlist id",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "format": "int64",
                    "type": "integer"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Matching track ids"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or invalid API key"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "No such playlist"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Evaluate the rules and return the matching track ids.",
        "tags": [
          "Playlists"
        ]
      }
    },
    "/api/playlists/{id}": {
      "delete": {
        "operationId": "delete",
        "parameters": [
          {
            "description": "Playlist id",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            },
            "description": "Deleted"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or invalid API key"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "No such playlist"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Delete a playlist and its entries.",
        "tags": [
          "Playlists"
        ]
      },
      "patch": {
        "operationId": "update",
        "parameters": [
          {
            "description": "Playlist id",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            },
            "description": "Updated"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or invalid API key"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "No such playlist"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Rename a playlist and/or set its icon.",
        "tags": [
          "Playlists"
        ]
      }
    },
    "/api/playlists/{id}/entries": {
      "get": {
        "operationId": "get_entries",
        "parameters": [
          {
            "description": "Playlist id",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PlaylistTrackEntry"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Entries, in playlist order"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or invalid API key"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "No such playlist"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Playlist entries, which carry their own id so the same track can appear twice.",
        "tags": [
          "Playlists"
        ]
      }
    },
    "/api/playlists/{id}/entries/{entry_id}": {
      "delete": {
        "operationId": "remove_entry",
        "parameters": [
          {
            "description": "Playlist id",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          },
          {
            "description": "Playlist entry id",
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            },
            "description": "Removed"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or invalid API key"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "No such playlist"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Remove one entry, leaving other copies of the same track in place.",
        "tags": [
          "Playlists"
        ]
      }
    },
    "/api/playlists/{id}/tracks": {
      "delete": {
        "operationId": "remove_tracks",
        "parameters": [
          {
            "description": "Playlist id",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrackIdsBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            },
            "description": "Removed"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or invalid API key"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "No such playlist"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Remove every entry for the given tracks.",
        "tags": [
          "Playlists"
        ]
      },
      "get": {
        "operationId": "get_playlist_tracks",
        "parameters": [
          {
            "description": "Playlist id",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "format": "int64",
                    "type": "integer"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Track ids, in playlist order"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or invalid API key"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "No such playlist"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Track ids in playlist order.",
        "tags": [
          "Playlists"
        ]
      },
      "post": {
        "operationId": "add_tracks",
        "parameters": [
          {
            "description": "Playlist id",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrackIdsBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            },
            "description": "Added"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or invalid API key"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "No such playlist"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Append tracks to a playlist.",
        "tags": [
          "Playlists"
        ]
      }
    },
    "/api/playlists/{id}/tracks/order": {
      "put": {
        "operationId": "reorder_tracks",
        "parameters": [
          {
            "description": "Playlist id",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReorderBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            },
            "description": "Reordered"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or invalid API key"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "No such playlist"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Reorder entries within a playlist.",
        "tags": [
          "Playlists"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
}